+++ client/lib/js/expressions.ts
@@ -32,8 +32,9 @@
  *       - `\'` - defines a single quote
  *       - `\n' - defines new line
  * - numerical values.
- * - numerical arrays: [x,y,z]
+ * - numerical arrays: [x,y,z].
  * - variables: numerical, textual, numerical arrays, objects.
+ * - variable array one-dimension indices.
  *
  * ## Built-in Variables
  *
@@ -65,6 +66,7 @@
  * `= iff(fps < 20, 'too few frames', 'lots of frames')`.
  * `=[2, 3] + [4, 5]`.
  * `=chart.labelsY.marginAfter`.
+ * `=foo[x-y+z]`.
  */
 namespace ABeamer {
 
@@ -379,6 +381,7 @@ namespace ABeamer {
         p.token.arrayValue = varValue as number[];
         p.token.sValue = undefined;
         p.token.numValue = undefined;
+
       } else if (expr[pos] === '.') {
         const varPropStart = ++pos;
         while (_isId(expr[pos], varPropStart === pos)) { pos++; }
@@ -387,7 +390,30 @@ namespace ABeamer {
           err(p, `Invalid object variable ${varName}`);
         }
         pos = _parseVars(p, varValue[varProp], varName + '.' + varProp, expr, pos);
+
+      } if (expr[pos] === '[') {
+        const varPropStart = ++pos;
+        let bracketCount = 1;
+        while (bracketCount > 0) {
+          switch (expr[pos]) {
+            case '[':
+              bracketCount++;
+              break;
+            case ']':
+              bracketCount--;
+              break;
+            case undefined:
+              err(p, `Invalid  variable indexing ${varName}`);
+          }
+          pos++;
+        }
+        const indexExpr = '=' + expr.substring(varPropStart, pos - 1);
+        const indexValue = calcExpr(indexExpr, p.args);
+        const arrayItem = (varValue as any[])[parseInt(indexValue as any)];
+        _parseVars(p, arrayItem, varName, indexExpr, 1);
+        pos++;
       }
+
     } else if (varTypeOf === 'boolean') {
       p.token.paType = ExFuncParamType.Number;
       p.token.numValue = (varValue as boolean) ? 1 : 0;